Remove specific requirement for SHA
One of the requirements stated a requirement for a matching SHA value. I took this to be more of a example of a mechanism since it did not specify which SHA algorithm. Replacing it with a requirement to be able to make a trust decision about the origin and integrity of the software.